+++ user_guide_src/source/changelog.rst
@@ -78,6 +78,7 @@ Release Date: Not Released
 
 -  :doc:`Database <database/index>` changes include:
 
+   -  Removed the option to disable the :doc:`Query Builder <database/query_builder>`.
    -  Removed driver-specific ``$curs_id`` property and ``get_cursor()``, ``stored_procedure()`` methods from OCI8 driver.
    -  Removed previously deprecated 'sqlite' driver (used for SQLite version 2; no longer shipped with PHP 5.4+).
    -  Removed method ``db_set_charset()`` and the ability to change a connection character set at runtime.